remove custom toolbar styling
Might be controversial, but my reasoning behind this is that some custom icons were not properly aligned due to inherent flaws in their design (a good example of this was with the undo/redo icons).

I also think keeping the design as close to vanilla as possible will better future proof the toolbar. I may revisit this if I can ever get in touch with the original creator.